Vector2 to avoid repeating vector2 in all names. Add SECs. Import unqualified operators.

+++ src/Math/Bezier.hs
@@ -9,7 +9,7 @@ bezier4 :: (RealFloat a) => a
         -> Vector2 a -> Vector2 a -> Vector2 a -> Vector2 a
         -> Vector2 a
 bezier4 mu v1 v2 v3 v4 = vsum (zipWith (*^) coeffs vectors)
-    where coeffs = binomialPoly 3 (1-mu) mu
+    where coeffs = binomialPoly (3::Integer) (1-mu) mu
           vectors = [v1,v2,v3,v4]
 
 --
